requirements/production.txt
alembic
bootstrap-flask
flask-sqlalchemy
flask-wtf
hypercorn
python-slugify
quart
SQLAlchemy
werkzeug
aiohttp